requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
joblib==1.0.0
lxml==4.6.2
numpy==1.19.4
pandas==1.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.5
requests==2.25.1
scikit-learn==0.24.0rc1
scipy==1.6.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.26.2
whitenoise==5.2.0
